requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg[binary]>=3.2.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
httpx>=0.27.0
python-multipart>=0.0.9
alembic>=1.13.1
langgraph>=0.2.0
langchain-core>=0.2.30
qdrant-client>=1.10.1
firebase-admin>=6.5.0
redis>=5.0.7
spacy>=3.7.5
sentence-transformers>=3.0.1
PyMuPDF>=1.24.7
python-docx>=1.1.2
reportlab>=4.2.2
python-dotenv>=1.0.1
tenacity>=8.4.2
